refactor: add view model unit test to view model module
