+++ viewmodel/src/test/java/com/baghdad/viewmodel/ActorGalleryViewModelTest.kt
@@ -1,6 +1,8 @@
-package com.baghdad.viewmodel.actorGallery
+package com.baghdad.viewmodel
 
 import com.baghdad.domain.usecase.actorDetails.GetActorGalleryUseCase
+import com.baghdad.viewmodel.actorGallery.ActorGalleryScreenEffect
+import com.baghdad.viewmodel.actorGallery.ActorGalleryViewModel
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.mockk
