--- viewmodel/src/test/java/com/baghdad/viewmodel/ExampleUnitTest.kt
@@ -1,16 +1,0 @@
-package com.baghdad.viewmodel
-
-import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.Assertions.assertEquals
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-class ExampleUnitTest {
-    @Test
-    fun addition_isCorrect() {
-        assertEquals(4, 2 + 2)
-    }
-}
